earning method update

+++ src/main/java/com/tanvir/features/liveroom/domain/valueobject/Earning.java
@@ -17,6 +17,6 @@ public class Earning {
     private String duration;
     private String durationString;
     private int validDays;
-    private int bonus;
+    private double bonus;
     private String bonusString;
 }
